Create Higher-Lower game Replit
This is a simple higher lower game in Python during my 100 days of Python Course. So basically it would prompt to compare statements, based on the popularity of  celebs on instagram [ followers] , it will check your answer and give you a feedback about your choice. Play around and try to score as high you can